requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-ollama==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.4.23
MarkupSafe==3.0.2
marshmallow==3.26.1
mercadopago==2.3.0
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
ollama==0.5.3
orjson==3.11.3
packaging==25.0
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.12.2
pydantic-settings==2.10.1
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zstandard==0.24.0
langchain-google-genai==2.1.12
fastapi-mail==1.5.7
APScheduler==3.10.0
Pillow
nsfw-image-detector==0.1.2
validate-docbr
pytz